Add qemu_xxx build products.
This patch adds 6 new build products that can be used to
generate a minimal system image, to be run under upstream QEMU
(with a few Android-related support patches). Note that this is
not the same as running under the Android emulator.

The goal is to use these system images to test upstream QEMU
changes. At a minimum, a modified QEMU should allow booting
the system images, provide a root shell, and let the guest
system visible / actionable through ADB.

The new products are:

  qemu_arm
  qemu_x86
  qemu_mips
  qemu_x86_64
  qemu_arm64
  qemu_mips64    [NOTE: Does not build at the moment].

To build one, do something like:

  . build/envsetup.sh
  lunch qemu_arm-userdebug
  make -j8

For now, assume the default goldfish kernel images are used
to boot.

Change-Id: I2c8caa27e6e40a2f86d0a05e039c89d30b8611b7